+++ maplibre_gl_platform_interface/lib/maplibre_gl_platform_interface.dart
@@ -14,6 +14,7 @@ part 'src/camera.dart';
 part 'src/circle.dart';
 part 'src/line.dart';
 part 'src/location.dart';
+part 'src/location_source.dart';
 part 'src/method_channel_maplibre_gl.dart';
 part 'src/symbol.dart';
 part 'src/fill.dart';